pyproject.toml
[project]
name = "watch-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.11",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=1.1.11",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
